dataset generation with themes

+++ gen_pdfs.sh
@@ -3,6 +3,8 @@
 
 TEMPLATES_DIR=/usr/share/haskell-pandoc/data/templates
 
+
+
 if test ! -d ${TEMPLATES_DIR}; then
 	echo No templates dir found
 	echo please download the TeX templates to ${TEMPLATES_DIR}
@@ -10,16 +12,21 @@ if test ! -d ${TEMPLATES_DIR}; then
 fi
 
 for i in *.md; do
-	echo processing ${i}
-	# generate PDF using base theme
-	pandoc --from=markdown --output=${i}-base.pdf ${i}
-	convert -density 200 ${i}-base.pdf ${i}-base.jpg
-	rm ${i}-base.pdf
-	# generate PDFs using different themes
-#	for j in ${TEMPLATES_DIR}/*.tex; do
-#		echo generating ${i}-${j%.*}.pdf
-#		pandoc --pdf-engine=xelatex --template=${j} --from=markdown --output=${i}-${j%.*}.pdf ${i}
-#		convert -density 200 ${i}-${j%.*}.pdf ${i}-${j%.*}.jpg
-#	done
+	if test ! -f ${i}-base.pdf; then
+		echo processing ${i}
+		# generate PDF using base theme
+		pandoc --pdf-engine=xelatex --template=./base.tex --from=markdown --output=${i}-base.pdf ${i}
+		convert -density 200 ${i}-base.pdf ${i}-base.jpg
+		rm ${i}-base.pdf
+		# generate PDFs using different themes
+		for j in ${TEMPLATES_DIR}/*.tex; do
+			j="$(basename ${j})"
+			j=${j%.*}
+			filename="${i}-${j}"
+			echo generating "${filename}.pdf"
+			./generate_file.sh "${filename}" "${i}" "${j}" &
+		done
+
+	fi
 done
